CSVs and Camera class
Environment contains information on each object, colour and coordinates
and radius. This information will be known to controller.

input CSV will contain basic information of what the camera would pass
to the controller. This includes the ID, and screen locations of the
box it puts around the object.

Need to work out how the camera will work out its height. Will have to
model the focal length of the camera somehow.