Controller improvements
Controller now works out the distance from object centre to the camera
centre in cm.

The angle between an object and camera is attempted although I do not
think that it is currently working.

Having issues to work out how the angle can be computed, especially if
you take into account the FOV changes of the camera.