Possible Distance calculation
Came up with an idea for calculating the distance…

Since we know the horizontal FOV of the camera and also the distance
covered per pixel we can work out the real distance covered by half of
the image.

Half of the image is 640px, so we can multiply this by the distance
covered per pixel and come up with a distance. Since we know the angle
to the edge of the screen is 75º we can work out a value for the height.

I’m not entirely sure if this is a correct solution but it does provide
a ratio which works in a way which you would expect when changing an
objects real radius but not changing its size in the picture (if the
real radius is smaller but is the same size on the image then we are
closer)