Changing the model
The way I was modelling the camera was slightly wrong. Instead of just
passing some pixel coordinates to a controller, the camera itself has
to model the fact if it can see an object or not.

To do this, the camera is created with a known coordinate X,Y and Z.
From here it then takes in the whole list of objects and goes through
them and works out if it would be able to see them at its given height.
If it would be able to see the object then it passes on the information
it normally would for the object to the controller.

Currently working out what information to pass - do I pass fake numbers
for now or do I try and set up the camera for real and see how it would
send data to a computer?