Clean up
Need to work on the camera a bit more. It must only pass the parameters
which the controller would be allowed to access i.e the id, width and
centre of an object. At the moment the controller ‘cheats’ by using the
distance calculated by the camera which isn’t how it would work in a
real model.