Method change
Basically this way of calculating the camera’s rotation above the plane
is wrong and won’t work. Although the value can be calculated, it
requires the use of variables which the controller would not have
access to.

It has access only to the x and y coordinates of the object relative to
its screen and the true x and y coordinates of the objects in the
environment.