Angle calculation
Currently trying to work out the angle of rotation by taking two points
and measuring the angle change between them and the Y axis.

I convert the known coordinates of the objects to what we would
estimate their position to be on the screen to do this. I don’t think
this is the best way of doing it but P3P is very difficult.