Angle Calculation
Think the angle calculation works for all points regardless of
translation.

Need to make sure it gives the correct rotation as sometimes it will
give ~30º if the rotation is 150º as it subtracts from 180º.

This should just be a few cases that need addressing. The model at this
stage should be pretty much complete.